refactored load classes into subdirectory and dedicated modules
