+++ src/sigmaepsilon/solid/fourier/loads/__init__.py
@@ -1,0 +1,6 @@
+from .loads import LoadCase, LoadGroup
+from .lineload import LineLoad
+from .pointload import PointLoad
+from .rectangleload import RectangleLoad
+
+__all__ = ["LoadCase", "LoadGroup", "LineLoad", "PointLoad", "RectangleLoad"]
